+++ ingestion/src/metadata/ingestion/source/dynamodb.py
@@ -24,7 +24,7 @@ from metadata.ingestion.ometa.ometa_api import OpenMetadata
 from metadata.ingestion.source.sql_source import SQLSourceStatus
 from metadata.utils.aws_client import AWSClient
 from metadata.utils.column_type_parser import ColumnTypeParser
-from metadata.utils.engines import get_connection, test_connection
+from metadata.utils.connections import get_connection, test_connection
 from metadata.utils.filters import filter_by_table
 
 logger: logging.Logger = logging.getLogger(__name__)
