+++ ingestion/src/metadata/ingestion/source/snowflake_usage.py
@@ -32,7 +32,7 @@ from metadata.ingestion.api.source import InvalidSourceException, Source, Source
 # This import verifies that the dependencies are available.
 from metadata.ingestion.models.table_queries import TableQuery
 from metadata.ingestion.source.sql_alchemy_helper import SQLSourceStatus
-from metadata.utils.engines import get_connection, test_connection
+from metadata.utils.connections import get_connection, test_connection
 from metadata.utils.helpers import get_start_and_end
 from metadata.utils.sql_queries import SNOWFLAKE_SQL_STATEMENT
