+++ ingestion/tests/unit/profiler/test_profiler.py
@@ -23,7 +23,7 @@ from metadata.orm_profiler.metrics.core import add_props
 from metadata.orm_profiler.metrics.registry import Metrics
 from metadata.orm_profiler.profiler.core import MissingMetricException, Profiler
 from metadata.orm_profiler.profiler.default import DefaultProfiler
-from metadata.utils.engines import create_and_bind_session
+from metadata.utils.connections import create_and_bind_session
 
 Base = declarative_base()
