+++ ingestion/tests/unit/profiler/test_runner.py
@@ -21,7 +21,7 @@ from sqlalchemy.orm import DeclarativeMeta, declarative_base
 
 from metadata.orm_profiler.profiler.runner import QueryRunner
 from metadata.orm_profiler.profiler.sampler import Sampler
-from metadata.utils.engines import create_and_bind_session
+from metadata.utils.connections import create_and_bind_session
 from metadata.utils.timeout import cls_timeout
 
 Base = declarative_base()
