+++ ingestion/tests/unit/profiler/test_sample.py
@@ -20,7 +20,7 @@ from sqlalchemy.orm import DeclarativeMeta, declarative_base
 from metadata.orm_profiler.metrics.registry import Metrics
 from metadata.orm_profiler.profiler.core import Profiler
 from metadata.orm_profiler.profiler.sampler import Sampler
-from metadata.utils.engines import create_and_bind_session
+from metadata.utils.connections import create_and_bind_session
 
 Base = declarative_base()
