+++ ingestion/tests/unit/profiler/test_session_validations.py
@@ -30,7 +30,7 @@ from metadata.generated.schema.tests.column.columnValuesToMatchRegex import (
     ColumnValuesToMatchRegex,
 )
 from metadata.orm_profiler.validations.core import validate
-from metadata.utils.engines import create_and_bind_session
+from metadata.utils.connections import create_and_bind_session
 
 EXECUTION_DATE = datetime.strptime("2021-07-03", "%Y-%m-%d")
 Base = declarative_base()
